+++ local/runner/opts.go
@@ -11,6 +11,6 @@ type Opts struct {
 	Users        int     `json:"users"`
 	Timeout      int     `json:"timeout"`
 	SuccessCodes []int   `json:"successCodes"`
-	Rate         float64 `json:"rate"`
+	Rate         float32 `json:"rate"`
 	// If Rate is 0, the requests will be made as fast as possible
 }
